# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "Health_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok",
                        "degraded"
                      ],
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up",
                          "responseTime": 12
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string",
                            "enum": [
                              "up",
                              "degraded",
                              "down"
                            ]
                          },
                          "responseTime": {
                            "type": "number",
                            "description": "Time the health indicator took to respond, in ms"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string",
                            "enum": [
                              "up",
                              "degraded",
                              "down"
                            ]
                          },
                          "responseTime": {
                            "type": "number",
                            "description": "Time the health indicator took to respond, in ms"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up",
                          "responseTime": 12
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string",
                            "enum": [
                              "up",
                              "degraded",
                              "down"
                            ]
                          },
                          "responseTime": {
                            "type": "number",
                            "description": "Time the health indicator took to respond, in ms"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "error",
                        "shutting_down"
                      ],
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up",
                          "responseTime": 12
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string",
                            "enum": [
                              "up",
                              "degraded",
                              "down"
                            ]
                          },
                          "responseTime": {
                            "type": "number",
                            "description": "Time the health indicator took to respond, in ms"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect",
                          "responseTime": 3005
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string",
                            "enum": [
                              "up",
                              "degraded",
                              "down"
                            ]
                          },
                          "responseTime": {
                            "type": "number",
                            "description": "Time the health indicator took to respond, in ms"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up",
                          "responseTime": 12
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect",
                          "responseTime": 3005
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string",
                            "enum": [
                              "up",
                              "degraded",
                              "down"
                            ]
                          },
                          "responseTime": {
                            "type": "number",
                            "description": "Time the health indicator took to respond, in ms"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "Auth_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationDto"
                }
              }
            }
          },
          "409": {
            "description": "EMAIL_TAKEN"
          }
        },
        "summary": "Sign up with email + password; a confirmation link is emailed",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "Auth_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_VERIFICATION_TOKEN"
          }
        },
        "summary": "Confirm the email address and sign in",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "operationId": "Auth_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Email a new confirmation link (always 202)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_CREDENTIALS"
          },
          "403": {
            "description": "EMAIL_NOT_VERIFIED"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/google": {
      "post": {
        "operationId": "Auth_google",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialSignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_ID_TOKEN"
          },
          "409": {
            "description": "EMAIL_TAKEN (email not verified by the provider)"
          }
        },
        "summary": "Sign in / sign up with a Google ID token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/apple": {
      "post": {
        "operationId": "Auth_apple",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialSignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_ID_TOKEN"
          },
          "409": {
            "description": "EMAIL_TAKEN (email not verified by the provider)"
          }
        },
        "summary": "Sign in / sign up with a Sign in with Apple identity token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "Auth_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_REFRESH_TOKEN"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "Auth_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "Users_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/roles": {
      "put": {
        "operationId": "Users_updateMyRoles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMyRolesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "Users_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admins only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "Tiralarc API",
    "description": "Public API consumed by the web app and the mobile apps.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SelfAssignableRole": {
        "type": "string",
        "enum": [
          "ARCHER",
          "COACH"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "displayName": {
            "type": "string",
            "maxLength": 100
          },
          "roles": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/SelfAssignableRole"
            }
          },
          "locale": {
            "type": "string",
            "enum": [
              "fr",
              "en"
            ],
            "description": "Language of the emails sent to the user."
          }
        },
        "required": [
          "email",
          "password",
          "roles"
        ]
      },
      "RegistrationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "verificationRequired": {
            "type": "boolean",
            "description": "Always true: a confirmation link was emailed; login is refused until it is used."
          }
        },
        "required": [
          "email",
          "verificationRequired"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token from the link emailed at registration."
          },
          "deviceName": {
            "type": "string",
            "description": "Human-readable device label (e.g. \"Pixel 9\", \"Firefox on Linux\"), shown in session lists.",
            "maxLength": 100
          }
        },
        "required": [
          "token"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Short-lived JWT, sent as `Authorization: Bearer <token>`."
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque, single-use token exchanged at /auth/refresh for a new pair."
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expiresIn": {
            "type": "number",
            "description": "Access token lifetime, in seconds."
          },
          "refreshExpiresIn": {
            "type": "number",
            "description": "Refresh token lifetime, in seconds."
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn",
          "refreshExpiresIn"
        ]
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "deviceName": {
            "type": "string",
            "description": "Human-readable device label (e.g. \"Pixel 9\", \"Firefox on Linux\"), shown in session lists.",
            "maxLength": 100
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SocialSignInDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "ID token returned by Google Identity Services / Sign in with Apple (web or native SDK)."
          },
          "nonce": {
            "type": "string",
            "description": "Nonce passed to the provider when requesting the token; checked if present."
          },
          "roles": {
            "type": "array",
            "description": "Roles for a new account. If omitted for a new account, it is created without roles and the client must ask the user (PUT /users/me/roles). Ignored for existing accounts.",
            "items": {
              "$ref": "#/components/schemas/SelfAssignableRole"
            }
          },
          "displayName": {
            "type": "string",
            "maxLength": 100,
            "description": "Apple only sends the name to the client, on first sign-in: forward it here."
          },
          "locale": {
            "type": "string",
            "enum": [
              "fr",
              "en"
            ]
          },
          "deviceName": {
            "type": "string",
            "description": "Human-readable device label (e.g. \"Pixel 9\", \"Firefox on Linux\"), shown in session lists.",
            "maxLength": 100
          }
        },
        "required": [
          "idToken"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "Role": {
        "type": "string",
        "enum": [
          "ADMIN",
          "ARCHER",
          "COACH"
        ],
        "description": "Empty for an account created through Google / Apple whose user has not chosen yet."
      },
      "AuthProvider": {
        "type": "string",
        "enum": [
          "GOOGLE",
          "APPLE"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "locale": {
            "type": "string",
            "example": "fr"
          },
          "roles": {
            "type": "array",
            "description": "Empty for an account created through Google / Apple whose user has not chosen yet.",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "hasPassword": {
            "type": "boolean",
            "description": "False for accounts created through Google / Apple only."
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthProvider"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "emailVerified",
          "displayName",
          "locale",
          "roles",
          "hasPassword",
          "providers",
          "createdAt"
        ]
      },
      "UpdateMyRolesDto": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "array",
            "minItems": 1,
            "description": "Replaces the self-assignable roles (ADMIN, if held, is kept).",
            "items": {
              "$ref": "#/components/schemas/SelfAssignableRole"
            }
          }
        },
        "required": [
          "roles"
        ]
      }
    }
  }
}